removed unecessary component
