+++ src/components/data.json
@@ -45,9 +45,7 @@
 		"text": "Ralonick Service Limited provides Steel structures and pipes fabrication & installation with expertise in construction capabilities and project management Ralonick delivers solutions by providing competent personnel that has the knowledge in all kinds of fabrication and installation works.",
 		"image2": "/images/services/steel2.png",
 		"image3": "",
-		"text2": "PIPING AND STRUCTURAL FABRICATION SERVICES: Pipe spool fabrication, Fabrication of boat landing & repairs, Structural Pipe Support, Steel Shelter Platform Fabrication, Pig Launcher fabrication",
-		"learn": "Learn More",
-		"btn": "/assets/learnMoreArrow.png"
+		"text2": "PIPING AND STRUCTURAL FABRICATION SERVICES: Pipe spool fabrication, Fabrication of boat landing & repairs, Structural Pipe Support, Steel Shelter Platform Fabrication, Pig Launcher fabrication"
 	},
 	{
 		"image": "https://media-exp1.licdn.com/dms/image/C4D12AQHrMuveWUWxwg/article-cover_image-shrink_600_2000/0/1584799834221?e=2147483647&v=beta&t=xIw3p9RkYq0nf6YZFNdwdAnAQFhZxkr_kDx7TjAa1xY",
