+++ src/pages/services/index.tsx
@@ -8,7 +8,7 @@ import { createSlug } from "../../utils/string.utils";
 import "swiper/css";
 import "swiper/css/navigation";
 import { Swiper, SwiperSlide } from "swiper/react";
-import { SwiperButtons } from "../../components/ServiceComp";
+import { SwiperButtons } from "..";
 
 const Services = () => {
 	const services = Data.map((service) => ({
