+++ src/styles/about.scss
@@ -26,7 +26,7 @@
 			gap: -4rem;
 			justify-content: flex-end;
 			@media screen and (min-width: $md) {
-				width: 30%;
+				width: 50%;
 			}
 		}
 	}
@@ -159,12 +159,10 @@
 	//===========END OF TRANSFORMING=============
 
 	.misson-vision-container {
-		border: 1px solid red;
 		.mission-vision-container_inner {
 			display: grid;
 			margin-top: 3.9375rem;
 			width: 100%;
-			border: 1px solid green;
 			grid-template-areas:
 				"a a a"
 				"b b b"
@@ -178,7 +176,7 @@
 			}
 
 			@media screen and (min-width: $md) {
-				width: 100%;
+				margin-inline: auto;
 				grid-template-areas:
 					"a a b"
 					"c . b"
@@ -187,12 +185,10 @@
 
 			.vision-frame {
 				grid-area: a;
-				width: 80%;
-				margin-inline: auto;
+				width: 100%;
 				@media screen and (min-width: $md) {
 					width: 100%;
 					height: 100%;
-					margin: 0;
 					img {
 						width: 100%;
 					}
@@ -204,16 +200,13 @@
 			}
 
 			.vision-description {
-				width: 80%;
-				margin-inline: auto;
+				width: 100%;
 				background: rgba(51, 51, 51, 0.02);
 				grid-area: b;
 				padding: 1.5rem 1.5rem;
 
 				@media screen and (min-width: $md) {
 					width: 100%;
-					height: 25rem;
-					margin: 0;
 				}
 				ul {
 					width: 100%;
@@ -228,13 +221,11 @@
 				background: rgba(51, 51, 51, 0.02);
 				grid-area: c;
 				margin: 1rem auto 0 auto;
-				width: 80%;
+				width: 100%;
 				padding: 1.5rem 1.5rem;
 				@media screen and (min-width: $md) {
-					width: 80%;
-					height: 25rem;
+					width: 100%;
 					margin: 0;
-					border: 3px solid greenyellow;
 				}
 				ul {
 					width: 100%;
@@ -248,13 +239,12 @@
 
 			.mission-frame {
 				grid-area: d;
-				width: 80%;
-				margin-inline: auto;
+				width: 100%;
+				// margin-inline: auto;
 				@media screen and (min-width: $md) {
 					width: 100%;
 					height: 100%;
 					margin: 0;
-					// margin-left: 1rem;
 				}
 
 				img {
