+++ src/styles/home.scss
@@ -322,7 +322,6 @@
 			gap: 2rem;
 			margin-top: 2rem;
 			width: 100%;
-			// height: 40rem;
 
 			@media screen and (max-width: 576px) {
 				width: 90vw;
@@ -342,8 +341,8 @@
 				}
 				.content {
 					display: flex;
-					// justify-content: space-between;
 					flex-direction: column;
+					justify-content: space-between;
 					.title {
 						font-size: 1rem;
 						color: $secondary;
@@ -360,9 +359,6 @@
 						p {
 							padding: 1.5rem 1.125rem;
 						}
-						// @media screen and (min-width: $md) {
-						// 	margin-top: 3rem;
-						// }
 					}
 				}
 			}
