+++ src/styles/services.scss
@@ -116,6 +116,7 @@
 									display: flex;
 									align-items: center;
 									gap: 2px;
+									text-decoration: none;
 									img {
 										width: 0.5rem;
 										aspect-ratio: 1/1;
